+++ android/app/src/main/java/org/mudamos/petition/firebase/Config.java
@@ -1,5 +1,5 @@
 package org.mudamos.petition.firebase;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.Map;
 import java.util.HashMap;