+++ src/utils.js
@@ -271,6 +271,8 @@ export const eligibleToSignPlip = ({ plip, user }) => {
 /* eslint-disable no-useless-escape */
 export const validateEmail = email => /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(email)
 
+export const isValidEmailRelaxed = test(/^.+@.+$/);
+
 export const validateCpf = cpf => {
   const cpfFormatted = unMaskCpf(cpf);
   return cpfFormatted.length === 11;